backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
pgvector==0.2.5
pydantic==2.7.1
groq==0.9.0
httpx==0.27.2
unidecode==1.3.8
pdfplumber==0.11.0
openpyxl==3.1.2
reportlab==4.2.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
